+++ src/library/js/pack.js
@@ -1,0 +1,26 @@
+/**
+ * Tag for template literals. Returns the packed element.
+ * @param  {Array} strings String part of template literal.
+ * @param  {...*} values Values part.
+ * @return {Object} Packed element.
+ */
+export default function pack (strings, ...values) {
+	const parts = [];
+	const valuesMap = {};
+	for (let index = 0; index < strings.length; index++) {
+		parts.push(strings[index]);
+		if (values.hasOwnProperty(index)) {
+			const anchor = getAnchor(valuesMap);
+			valuesMap[anchor] = values[index];
+			parts.push(anchor);
+		}
+	}
+	const htmlString = parts.join('').trim();
+	return { htmlString, valuesMap };
+}
+
+function getAnchor (valuesMap) {
+	const length = Object.keys(valuesMap).length;
+	const anchor = `{%${length}%}`;
+	return anchor;
+}
